Refactor iOS e2e tests
