+++ example/__tests__/setup/index.js
@@ -1,3 +1,1 @@
 import './getLabel'
-import './loginToTwitter'
-import './loginToTwitterWithSystemAccount'
